requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
azure-storage-blob==12.17.0
python-dotenv==1.0.0
torch==2.1.0
torchvision==0.16.0
pydantic==2.5.2
pillow 